+++ compiler/damlc/daml-doc/src/DA/Daml/Doc/Transform/Annotations.hs
@@ -1,18 +1,22 @@
 -- Copyright (c) 2020 The DAML Authors. All rights reserved.
 -- SPDX-License-Identifier: Apache-2.0
 
+module DA.Daml.Doc.Transform.Annotations
+    ( applyAnnotations
+    ) where
 
-module DA.Daml.Doc.Annotate(applyAnnotations) where
+import DA.Daml.Doc.Types
 
-import           DA.Daml.Doc.Types
 import qualified Data.Text as T
-import           Data.List.Extra
+import Data.List.Extra
+import Control.Applicative ((<|>))
 
-
--- | Apply the annotation HIDE to hide either modules or declarations
+-- | Apply HIDE and MOVE annotations.
 applyAnnotations :: [ModuleDoc] -> [ModuleDoc]
 applyAnnotations = applyMove . applyHide
 
+-- | Apply the MOVE annotation, which moves all the docs from one
+-- module to another.
 applyMove :: [ModuleDoc] -> [ModuleDoc]
 applyMove = map (foldr1 g) . groupSortOn (modulePriorityKey . md_name) . map f
     where
@@ -23,7 +27,7 @@ applyMove = map (foldr1 g) . groupSortOn (modulePriorityKey . md_name) . map f
         g m1 m2 = ModuleDoc
             { md_anchor = md_anchor m1
             , md_name = md_name m1
-            , md_descr = md_descr m1
+            , md_descr = md_descr m1 <|> md_descr m2
             , md_adts = md_adts m1 ++ md_adts m2
             , md_functions = md_functions m1 ++ md_functions m2
             , md_templates = md_templates m2 ++ md_templates m2
@@ -35,6 +39,9 @@ applyMove = map (foldr1 g) . groupSortOn (modulePriorityKey . md_name) . map f
         modulePriorityKey :: Modulename -> (Int,Modulename)
         modulePriorityKey m = (if m == "Prelude" then 0 else 1, m)
 
+-- | Apply the HIDE annotation, which removes the current subtree from
+-- the docs. This can be applied to an entire module, or to a specific
+-- type, a constructor, a field, a class, a method, or a function.
 applyHide :: [ModuleDoc] -> [ModuleDoc]
 applyHide = concatMap onModule
     where
